requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTX==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
black==23.12.0
